cleaned up calc and p_invertor
moved a lot of testing/now obselete old versions to TEST_CALC.py, will be removed eventually.
